Add issue # to TODOs, custom decorator to extract publickey from JWT, replaces workgroup agent calls to retrieve bpiSubjectById with corresponding bpiSubjectAgents
