+++ examples/bri-3/src/bri/workgroup/workgroups/agents/workgroups.agent.ts
@@ -5,53 +5,13 @@ import { Workstep } from '../../worksteps/models/workstep';
 
 import { uuid } from 'uuidv4';
 import { Workgroup } from '../models/workgroup';
-import { BpiSubjectStorageAgent } from '../../../identity/bpiSubjects/agents/bpiSubjectsStorage.agent';
 import { WorkgroupStorageAgent } from './workgroupStorage.agent';
 import { WORKGROUP_NOT_FOUND_ERR_MESSAGE } from '../api/err.messages';
-import { NOT_FOUND_ERR_MESSAGE as BPISUBJECT_NOT_FOUND_ERR_MESSAGE } from '../../../identity/bpiSubjects/api/err.messages';
 
 // Agent methods have extremely declarative names and perform a single task
 @Injectable()
 export class WorkgroupAgent {
-  constructor(
-    private workgroupStorageAgent: WorkgroupStorageAgent,
-    private bpiSubjectStorageAgent: BpiSubjectStorageAgent,
-  ) {}
-
-  public async fetchBpiSubjectsByIdAndThrowIfNoneExist(
-    bpiSubjectIds: string[],
-  ): Promise<BpiSubject[]> {
-    const bpiSubjects = await this.bpiSubjectStorageAgent.getBpiSubjectsById(
-      bpiSubjectIds,
-    );
-
-    if (bpiSubjectIds.length !== bpiSubjects.length) {
-      throw new NotFoundException(BPISUBJECT_NOT_FOUND_ERR_MESSAGE);
-    }
-
-    bpiSubjects.forEach((bp) => {
-      if (!bpiSubjectIds.includes(bp.id)) {
-        throw new NotFoundException(BPISUBJECT_NOT_FOUND_ERR_MESSAGE);
-      }
-    });
-
-    return bpiSubjects;
-  }
-
-  public async fetchBpiSubjectByPublicKeyAndThrowIfNoneExist(
-    bpiSubjectPublicKey: string,
-  ): Promise<BpiSubject> {
-    const bpiSubject =
-      await this.bpiSubjectStorageAgent.getBpiSubjectByPublicKey(
-        bpiSubjectPublicKey,
-      );
-
-    if (!bpiSubject) {
-      throw new NotFoundException(BPISUBJECT_NOT_FOUND_ERR_MESSAGE);
-    }
-
-    return bpiSubject;
-  }
+  constructor(private workgroupStorageAgent: WorkgroupStorageAgent) {}
 
   public createNewWorkgroup(
     name: string,
